rename currency utils
