+++ src/bot/lib/utils/currency.js
